requirements.txt
Django==2.2.6
django-admin-sortable2==0.7.6
django-tinymce==3.0.2
Pillow==7.1.2
python-dotenv==0.13.0
requests==2.24.0